extract product description to its own component
This is a small step towards reducing the complexity of the ProductSection component. If we wanted to, we can change the Accordion component for a different type of component without affecting the entire ProductDescription component. This will also
make it easier to read the layout of the ProductSection component.
