+++ frontend/templates/product/sections/ProductSection/index.tsx
@@ -16,15 +16,10 @@ import {
    Link,
    StackProps,
    Text,
-   ThemeTypings,
    VStack,
 } from '@chakra-ui/react';
 import { CompatibleDevice } from '@components/common';
-import {
-   faCircleExclamation,
-   faCircleInfo,
-   faTriangleExclamation,
-} from '@fortawesome/pro-solid-svg-icons';
+import { faCircleExclamation } from '@fortawesome/pro-solid-svg-icons';
 import { useAppContext } from '@ifixit/app';
 import { isLifetimeWarranty } from '@ifixit/helpers';
 import { FaIcon } from '@ifixit/icons';
@@ -43,6 +38,7 @@ import { ProductVideos } from './ProductVideos';
 import { Prop65Warning } from './Prop65Warning';
 import { useInternationalBuyBox } from '@templates/product/hooks/useInternationalBuyBox';
 import { InternationalBuyBox } from './InternationalBuyBox';
+import { ProductDescription } from './ProductDescription';
 
 export type ProductSectionProps = {
    product: Product;
@@ -190,73 +186,14 @@ export function ProductSection({
                      <CustomAccordionButton>Description</CustomAccordionButton>
                      <CustomAccordionPanel>
                         <VStack>
-                           <VariantDescription>
-                              {selectedVariant.description ??
-                                 product.descriptionHtml}
-                           </VariantDescription>
-                           {selectedVariant.note && (
-                              <Alert
-                                 status="info"
-                                 borderWidth={1}
-                                 borderColor="brand.300"
-                                 borderRadius="md"
-                                 alignItems="flex-start"
-                              >
-                                 <FaIcon
-                                    icon={faCircleInfo}
-                                    h="4"
-                                    mt="0.5"
-                                    mr="2.5"
-                                    color="brand.500"
-                                 />
-                                 <AlertText colorScheme="brand">
-                                    {selectedVariant.note}
-                                 </AlertText>
-                              </Alert>
-                           )}
-                           {selectedVariant.disclaimer && (
-                              <Alert
-                                 status="warning"
-                                 borderWidth={1}
-                                 borderColor="orange.300"
-                                 borderRadius="md"
-                                 alignItems="flex-start"
-                              >
-                                 <FaIcon
-                                    icon={faCircleExclamation}
-                                    h="4"
-                                    mt="0.5"
-                                    mr="2.5"
-                                    color="orange.500"
-                                 />
-                                 <AlertText colorScheme="orange">
-                                    {selectedVariant.disclaimer}
-                                 </AlertText>
-                              </Alert>
-                           )}
-                           {selectedVariant.warning && (
-                              <Alert
-                                 status="error"
-                                 borderWidth={1}
-                                 borderColor="red.300"
-                                 borderRadius="md"
-                                 alignItems="flex-start"
-                              >
-                                 <FaIcon
-                                    icon={faTriangleExclamation}
-                                    color="red.500"
-                                    h="4"
-                                    mt="0.5"
-                                    mr="2.5"
-                                 />
-                                 <AlertText colorScheme="red">
-                                    {selectedVariant.warning}
-                                 </AlertText>
-                              </Alert>
-                           )}
+                           <ProductDescription
+                              product={product}
+                              selectedVariant={selectedVariant}
+                           />
                         </VStack>
                      </CustomAccordionPanel>
                   </AccordionItem>
+
                   <WikiHtmlAccordianItem title="Kit contents">
                      {selectedVariant.kitContents}
                   </WikiHtmlAccordianItem>
@@ -478,68 +415,6 @@ function NotForSaleAlert(props: AlertProps) {
             </p>
          </Box>
       </Alert>
-   );
-}
-
-type VariantDescriptionProps = {
-   children: string;
-};
-
-function VariantDescription({ children }: VariantDescriptionProps) {
-   return (
-      <Box
-         dangerouslySetInnerHTML={{
-            __html: children,
-         }}
-         fontSize="sm"
-         sx={{
-            ul: {
-               my: 3,
-               pl: 5,
-            },
-            p: {
-               mb: 3,
-               _last: {
-                  mb: 0,
-               },
-            },
-            a: {
-               color: 'brand.500',
-            },
-            'a:hover': {
-               textDecoration: 'underline',
-            },
-         }}
-      />
-   );
-}
-
-type AlertTextProps = {
-   children: string;
-   colorScheme: ThemeTypings['colorSchemes'];
-};
-
-function AlertText({ children, colorScheme }: AlertTextProps) {
-   return (
-      <Box
-         fontSize="sm"
-         sx={{
-            a: {
-               fontWeight: 'bold',
-               textDecoration: 'underline',
-               transition: 'all 300ms',
-            },
-            'a:hover': {
-               color: `${colorScheme}.800`,
-            },
-            'p:not(:first-of-type)': {
-               mt: 2,
-            },
-         }}
-         dangerouslySetInnerHTML={{
-            __html: children,
-         }}
-      />
    );
 }
